Alter how the font is set
Changing from 8pt down to 7pt produces a much larger decrease in size
then 9 -> 8 so I wondered if I was setting the value wrong. Turns out
I wasn't so something else is going on. This new way of setting seems
to be the correct way as this is how it is suggested to be done the
emacs wiki so lets stick with it.

Also alter how we call the doom-mode line as start up is starting to
get a big sliggish.